Make SRTPCipherCTR an interface
