+++ src/org/jitsi/impl/neomedia/transform/srtp/BaseSRTPCryptoContext.java
@@ -186,13 +186,13 @@ class BaseSRTPCryptoContext
         switch (policy.getEncType())
         {
         case SRTPPolicy.AESCM_ENCRYPTION:
-            cipherCtr = new SRTPCipherCTR(AES.createBlockCipher());
+            cipherCtr = new SRTPCipherCTRJava(AES.createBlockCipher());
             encKey = new byte[encKeyLength];
             saltKey = new byte[saltKeyLength];
             break;
 
         case SRTPPolicy.TWOFISH_ENCRYPTION:
-            cipherCtr = new SRTPCipherCTR(new TwofishEngine());
+            cipherCtr = new SRTPCipherCTRJava(new TwofishEngine());
             encKey = new byte[encKeyLength];
             saltKey = new byte[saltKeyLength];
             break;
